Fixes to support innerspace injection
IS mucks with things a bit, so we need to adapt. In particular,
the result of getting the window proc is garbage, so I need to
go detect it the old fashioned way.

Secondly, the graphics device may be modified, so I need to make
sure that I'm always checking the graphics device that EQ has
stored. This might be more important if the plugin is loaded at
startup, and might actually fix other issues, too.