Feature: filter geometry from scene via MaxZoneExtents
Add new structure of maximum extents for geometry that should
be loaded per zone. If geometry falls outside of the extents
specified in the MaxZoneExtents map, it will be filtered out.

This fixes issues in certain zones that have garbage geometry
far outside of the boundaries of the normal zone area. This
causes problems especially if the geometry is far below the
mesh, due to maximum height issues with recast.