Add command line arguments to temporary adjust log level (#49).
All commands now accept log=<level> which will modify the log output for
the full duration of the command. If the command starts a navigation,
then the log level will be modified for the duration of the navigation
as well.

Valid levels, in order of increasing priority:
* trace
* debug
* info
* warning
* error
* critical
* off

For example, the following command normally will print an error if no
path is active:

  /nav stop

But by changing the log level to 'off', then nothing will be printed:

  /nav stop log=off