Fix memory allocation issues by pre-calculating decompressed buffer size
Introduce v5 mesh header which adds the following fields to the mesh file:
- header size
- decompressed size

The header size is used for forwards compatibility, in case we want to add
additional optional fields to the header in the future.

The decompressed size is so that we know how big the mesh data is. If we
don't have the decompressed size (v4 header) then we will decompress the
data twice - once to determine the size of the buffer.

This is all to fix the large number of allocations that are occurring in
the buffer because we don't know how large it should be.